+++ app/src/main/java/com/example/tutorialjetpack/domain/model/UserModel.kt
@@ -4,7 +4,7 @@ data class UserModel(
     val name: String,
     val age: Int,
     val gender: String,
-    val height: Int,
-    val weight: Int,
+    val height: Double,
+    val weight: Double,
 )
 //error commit
